+++ packages/client/src/index.ts
@@ -22,11 +22,11 @@ import glspEditLabelModule from './features/edit-label/di.config';
 import modelHintsModule from './features/hints/di.config';
 import glspHoverModule from './features/hover/di.config';
 import layoutModule from './features/layout/di.config';
-import modelSourceWatcherModule from './features/model-source-watcher/di.config';
 import glspMouseToolModule from './features/mouse-tool/di.config';
 import { navigationModule } from './features/navigation/di.config';
 import saveModule from './features/save/di.config';
 import glspSelectModule from './features/select/di.config';
+import sourceModelWatcherModule from './features/source-model-watcher/di.config';
 import toolFeedbackModule from './features/tool-feedback/di.config';
 import paletteModule from './features/tool-palette/di.config';
 import { enableDefaultToolsOnFocusLossModule, toolsModule } from './features/tools/di.config';
@@ -66,7 +66,6 @@ export * from './features/hints/model';
 export * from './features/hints/type-hints';
 export * from './features/hover/hover';
 export * from './features/layout/layout-elements-action';
-export * from './features/model-source-watcher/model-source-changed-action-handler';
 export * from './features/mouse-tool/mouse-tool';
 export * from './features/navigation/navigation-action-handler';
 export * from './features/navigation/navigation-target-resolver';
@@ -74,6 +73,7 @@ export * from './features/rank/model';
 export * from './features/reconnect/model';
 export * from './features/save/model';
 export * from './features/save/save-keylistener';
+export * from './features/source-model-watcher/source-model-changed-action-handler';
 export * from './features/tool-feedback/change-bounds-tool-feedback';
 export * from './features/tool-feedback/creation-tool-feedback';
 export * from './features/tool-feedback/css-feedback';
@@ -121,6 +121,6 @@ export {
     navigationModule,
     markerNavigatorModule,
     glspDecorationModule,
-    modelSourceWatcherModule,
+    sourceModelWatcherModule as modelSourceWatcherModule,
     glspViewportModule
 };